undo moving command specific files
